pvcracks/vae/requirements_vae.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cudatoolkit=11.8.0=h6a678d5_0
ipykernel=5.5.5=py39hef51801_0
ipython=8.16.1=pyh0d859eb_0
ipython_genutils=0.2.0=py_1
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hf3d152e_0
jupyter_server=1.24.0=pyhd8ed1ab_0
jupyterlab=3.5.3=pyhd8ed1ab_0
jupyterlab-variableinspector=3.0.9=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.22.0=py39h06a4308_0
matplotlib=3.6.2=py39h06a4308_0
matplotlib-base=3.6.2=py39h945d387_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
notebook=6.5.6=pyha770c72_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.26.2=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.76=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=py39h06a4308_0
pandas=1.5.2=py39h417a72b_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.2.0=pyhd3eb1b0_1
patsy=0.5.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py39h6a678d5_0
pip=23.2.1=py39h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
plotly=5.18.0=pypi_0
python=3.9.5=h12debd9_4
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.18.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=2.0.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=py39h6a678d5_1
scikit-learn=1.2.0=py39h6a678d5_0
scipy=1.10.1=py39h14f4228_0
seaborn=0.12.2=py39h06a4308_0
termcolor=2.3.0=pyhd8ed1ab_0
torch-geometric=2.4.0=pypi_0
torchaudio=2.0.0=py39_cu118
torchinfo=1.8.0=pyhd8ed1ab_0
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu118
xgboost=2.0.2=pypi_0
yaml=0.2.5=h7b6447c_0

